crypt: move encryption to xts module
